% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/installOffPiste.R
\name{installOffPiste}
\alias{installOffPiste}
\title{Use install.packages with specifically tailored `repos` parameter.}
\usage{
installOffPiste(x, version = "3.18", quiet = TRUE, ...)
}
\arguments{
\item{x}{character vector of packages to be installed}

\item{version}{character(1) defaults to "3.18"}

\item{quiet}{logical(1) if TRUE (default), suppress warnings about the search for binaries}

\item{\dots}{arguments passed to `install.packages`, exclusive of `pkgs` and `repos`}
}
\value{
result of `install.packages`
}
\description{
Use install.packages with specifically tailored `repos` parameter.
}
